requirements.txt
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
certifi==2022.12.7
charset-normalizer==3.0.1
hydra-core==1.3.1
idna==3.4
imageio==2.24.0
joblib==1.2.0
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-python-headless==4.7.0.68
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
psutil==5.9.4
pydicom==2.3.1
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.3
pylibjpeg-openjpeg==1.3.1
python-dateutil==2.8.2
python-gdcm==3.0.20
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.11.0
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
